JD Resume Analyser/requirements.txt
streamlit
python-dotenv
langchain-groq
pdfplumber
pdf2image
pytesseract
pillow
python-docx
